% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchVotacaoPlenarioCamara.R
\name{buildRollcallDataset}
\alias{buildRollcallDataset}
\title{Fetch bills discussed and voted on the plenary}
\usage{
buildRollcallDataset(
  year = 2020,
  type = "",
  ascii = FALSE,
  outfile = "cham_data_votes/data_"
)
}
\arguments{
\item{year}{an itneger the year querying}

\item{type}{a string for querying a particular type of bill}

\item{ascii}{a logical}

\item{outfile}{folder and name for the file}
}
\description{
Fetch bills discussed and voted on the plenary
}
\examples{
# data <- buildRollcallDataset(year=2020)

}
