% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVotacaoCamara.R
\name{fetchVotacoesCamara}
\alias{fetchVotacoesCamara}
\title{Fetch deputie's roll-call vote given a roll-call id}
\usage{
fetchVotacoesCamara(id)
}
\arguments{
\item{id}{A string for the rollcall id}
}
\description{
Fetch deputie's roll-call vote given a roll-call id
}
\examples{
 dat <- fetchVotacoesCamara("2206395-58")

}
