% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVotacaoCamara.R
\name{fetchVotacoesOrientacoesCamara}
\alias{fetchVotacoesOrientacoesCamara}
\title{Fetch roll-call party/bloc vote orientation}
\usage{
fetchVotacoesOrientacoesCamara(id)
}
\arguments{
\item{id}{A string for the roll-call id}
}
\description{
Fetch roll-call party/bloc vote orientation
}
\examples{
\dontrun{
 dat <- fetchVotacoesOrientacoesCamara("2206395-58")
 }
 
}
