% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listaDeputadosAtivos.R
\name{listaDeputadosAtivos}
\alias{listaDeputadosAtivos}
\title{Fetch a list of actual deputies}
\usage{
listaDeputadosAtivos(
  url = "https://dadosabertos.camara.leg.br/api/v2/deputados"
)
}
\arguments{
\item{url}{the web address}
}
\description{
Fetch a list of actual deputies
}
