% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVotacaoCamara.R
\name{loadCamaraProposals}
\alias{loadCamaraProposals}
\title{Load deputie's roll-call votes}
\usage{
loadCamaraProposals(year)
}
\arguments{
\item{year}{An integer}
}
\description{
Load deputie's roll-call votes
}
\examples{

# dados1 <- loadCamaraVotes(2020)

}
