% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVotacaoCamara.R
\name{loadVotacoesCamara}
\alias{loadVotacoesCamara}
\title{Load deputie's roll-call votes}
\usage{
loadVotacoesCamara(year)
}
\arguments{
\item{year}{An integer}
}
\description{
Load deputie's roll-call votes
}
\examples{
\dontrun{
# dados3 <- loadVotacoesCamara(2020)
}
}
