% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVotacaoCamara.R
\name{loadVotacoesOrientacoesCamara}
\alias{loadVotacoesOrientacoesCamara}
\title{Load orietation of deputie's roll-call votes}
\usage{
loadVotacoesOrientacoesCamara(year)
}
\arguments{
\item{year}{An integer}
}
\description{
Load orietation of deputie's roll-call votes
}
\examples{
\dontrun{

# dados2 <- loadVotacoesOrientacoesCamara(2020)
}

}
