% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{A dataframe of Brazilian Chamber proposals.}
\format{
A data frame with 52049 rows and 5 variables
}
\usage{
proposals
}
\description{
\itemize{
  \item bill_id \code{character}. Acronyms of Brazilian states.
  \item type_bill \code{character}. Full names of Brazilian states.
  \item year_bill \code{character}. Full names of Brazilian states.
  \item rollcall_subject \code{character}. Full names of Brazilian states.
  \item rollcall_keywords \code{character}. Full names of Brazilian states.
}
}
\keyword{datasets}
