% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRollcallDataset.R
\name{buildRollcallDataset}
\alias{buildRollcallDataset}
\title{Fetch bills discussed and voted on the plenary}
\usage{
buildRollcallDataset(
  year = 2021,
  type = "",
  download = TRUE,
  ascii = FALSE,
  filter = TRUE
)
}
\arguments{
\item{year}{an itneger the year querying}

\item{type}{a string for querying a particular type of bill}

\item{download}{a logical, if TRUE it will download a database of bills before combining all data}

\item{ascii}{a logical}

\item{filter}{a logical, if TRUE, it will filter data to a few types of bills, such as PEC, PL, PLP, PDL, PDC, MPV etc}
}
\description{
Fetch bills discussed and voted on the plenary
}
\examples{
data <- buildRollcallDataset(year = 2021)

}
