% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformVotes.R
\name{transformVotes}
\alias{transformVotes}
\title{Transform the database of rollcall votes}
\usage{
transformVotes(.data, filter = FALSE)
}
\arguments{
\item{.data}{the data frame}

\item{filter}{a logical, if TRUE, only rollcall votes with government orientation is returned. So "liberado" or missing cases are filtered from data output.}
}
\description{
Transform a database of rollcall votes, use only after downloading data with `buildRollcallDataset`
}
